refactor before playback support
The goal of this work is to make it easier to make a  copy of the Rete instance for use by the playback.

- make NodeEditorMST just include the methods that are overriding the Rete NodeEditor methods.
- Create ReteManager class that has the extra methods from NodeEditor along with methods from DataflowProgram which don't need to be in the component.
- remove some handlers from DataflowTile they can be implemented in the ReteManager or DataflowProgram directly
