+++ src/plugins/dataflow/components/ui/dataflow-drop-zone.tsx
@@ -3,19 +3,19 @@ import { DragEndEvent, useDndMonitor, useDroppable } from "@dnd-kit/core";
 import { observer } from "mobx-react";
 
 import { dataflowDroppableId, getNodeType, isNodeDraggableId } from "../dataflow-types";
-import { NodeEditorMST } from "../../nodes/node-editor-mst";
+import { ReteManager } from "../../nodes/rete-manager";
 
 interface IDataflowDropZoneProps {
   addNode: (nodeType: string, position?: [number, number]) => void;
   children?: any;
   className?: string;
-  programEditor: NodeEditorMST;
+  reteManager?: ReteManager;
   readOnly?: boolean;
   style?: any;
   tileId: string;
 }
 export const DataflowDropZone = observer((
-  { addNode, children, className, programEditor, readOnly, style, tileId }: IDataflowDropZoneProps
+  { addNode, children, className, reteManager, readOnly, style, tileId }: IDataflowDropZoneProps
 ) => {
 
   const droppableId = dataflowDroppableId(tileId);
@@ -26,7 +26,7 @@ export const DataflowDropZone = observer((
   };
   useDndMonitor({
     onDragEnd: (event: DragEndEvent) => {
-      if (readOnly) return;
+      if (readOnly || !reteManager) return;
       const draggableId = event.active.id.toString();
       if (event.over?.id === droppableId && isNodeDraggableId(draggableId)) {
         const nodeType = getNodeType(draggableId);
@@ -34,7 +34,7 @@ export const DataflowDropZone = observer((
 
         const clientX = pointerEvent.clientX + event.delta.x;
         const clientY = pointerEvent.clientY + event.delta.y;
-        const reteArea = programEditor.area.area;
+        const reteArea = reteManager.area.area;
 
         // This was taken from Rete's Area.setPointerFrom
         const { x, y } = reteArea.content.getPointerFrom({clientX, clientY} as MouseEvent);
