+++ src/plugins/dataflow/nodes/node-editor-mst.ts
@@ -1,283 +1,21 @@
 import { NodeEditor } from "rete";
-import { DataflowEngine } from "rete-engine";
-import { structures } from "rete-structures";
-import { onPatch } from "mobx-state-tree";
-import { AreaExtra, Schemes } from "./rete-scheme";
+import { Schemes } from "./rete-scheme";
 import {
   DataflowNodeModel, IDataflowNodeModel, DataflowProgramModelType, ConnectionModel
 } from "../model/dataflow-program-model";
-import { NumberNode } from "./number-node";
-import { MathNode } from "./math-node";
-import { CounterNode } from "./counter-node";
-import { LogicNode } from "./logic-node";
-import { GeneratorNode } from "./generator-node";
-import { IBaseNode, IBaseNodeModel, NodeClass } from "./base-node";
-import { uniqueId } from "../../../utilities/js-utils";
-import { INodeServices } from "./service-types";
-import { LogEventName } from "../../../lib/logger-types";
-import { logTileChangeEvent } from "../../../models/tiles/log/log-tile-change-event";
-import { AreaExtensions, AreaPlugin } from "rete-area-plugin";
-import { DemoOutputNode } from "./demo-output-node";
-import { LiveOutputNode } from "./live-output-node";
-import { DataflowContentModelType } from "../model/dataflow-content";
-import { IStores } from "../../../models/stores/stores";
-import { SensorNode } from "./sensor-node";
-import { TransformNode } from "./transform-node";
-import { TimerNode } from "./timer-node";
-import { ControlNode } from "./control-node";
-import { getNewNodePosition } from "./utilities/view-utilities";
+import { IBaseNode, IBaseNodeModel } from "./base-node";
 
-export class NodeEditorMST extends NodeEditor<Schemes> implements INodeServices {
+export class NodeEditorMST extends NodeEditor<Schemes> {
   public reteNodesMap: Record<string, Schemes['Node']> = {};
-
-  public engine = new DataflowEngine<Schemes>();
-  public area: AreaPlugin<Schemes, AreaExtra>;
 
   constructor(
     private mstProgram: DataflowProgramModelType,
-    private tileId: string,
-    private div: HTMLElement,
-    private mstContent: DataflowContentModelType,
-    public stores: IStores,
-    public runnable: boolean | undefined,
-    public readOnly: boolean | undefined,
+    private process: () => void,
+    private createReteNodeFromNodeModel: (id: string, model: IBaseNodeModel) => IBaseNode | undefined
   ) {
     super();
-
-    this.use(this.engine);
-
-    this.area = new AreaPlugin<Schemes, AreaExtra>(div);
-
-    // Disable the zoom handler which zooms on wheel and double click
-    this.area.area.setZoomHandler(null);
-
-    AreaExtensions.selectableNodes(this.area, AreaExtensions.selector(), {
-      accumulating: AreaExtensions.accumulateOnCtrl()
-    });
-
-    this.area.addPipe((context) => {
-      if (context.type !== "nodedragged") return context;
-
-      const id = context.data.id;
-      const nodeView = this.area.nodeViews.get(id);
-
-      // This should not happen, but it is possible in theory
-      if (!nodeView) return context;
-
-      const nodeModel = mstProgram.nodes.get(id);
-
-      // This should also not happen but seems more likely
-      if (!nodeModel) {
-        console.warn("Cannot find MST node model for a dragged Rete node");
-        return context;
-      }
-
-      nodeModel.setPosition(nodeView.position);
-      return context;
-    });
-
-    // Useful for debugging:
-    // this.area.addPipe((context) => {
-    //   if (["pointermove", "render", "rendered"].includes(context.type)) return context;
-
-    //   console.warn("area event", context, this.area.area.transform);
-    //   return context;
-    // });
-
-    // onPatch(mstProgram.nodes, (patch, reversePatch) => {
-    //   console.log("nodes patch", patch);
-    //   if (patch.op === "replace") {
-    //     // don't do anything with this yet, but it might be needed
-    //     // to support applySnapshot
-    //     return;
-    //   }
-    //   // Using a method to parse the path would be better here
-    //   const pattern = /^\/([^/[]*)$/;
-    //   const result = patch.path.match(pattern);
-    //   if (!result) {
-    //     return;
-    //   }
-    //   const id = result[1];
-    //   if (patch.op === "add") {
-    //     const node = this.getNode(id);
-    //     if (!node) return;
-    //     this.emit({ type: 'nodecreated', data: node });
-    //   }
-
-    //   if (patch.op === "remove") {
-    //     // I'm not sure what the value of patch will be here. And it is going to be tricky
-    //     // to instantiate a real node back from the patch information. I'm hoping that
-    //     // stuff that pays attention to this just looks at the id, but we'll see.
-    //     this.emit({ type: 'noderemoved', data: { id } as any});
-    //   }
-    // });
-
-    onPatch(mstProgram.connections, patch => {
-      console.log("connections patch", patch);
-
-      // If this is a connection being added:
-      // we need to send the connection instance not data
-      // this.emit({ type: 'connectioncreated', data })
-
-      // If this ia connection being removed
-      // we will no longer have the real connection instance to pass
-      // but we can probably send previous snapshot somehow
-      // this.emit({ type: 'connectionremoved', data: connection })
-    });
   }
 
-  public async notifyAboutExistingObjects() {
-    for (const node of this.getNodes()) {
-      // We need to wait for this before we add connections and translate the nodes
-      await this.emit({ type: 'nodecreated', data: node });
-    }
-    for (const connection of this.getConnections()) {
-      await this.emit({ type: 'connectioncreated', data: connection});
-    }
-
-    // FIXME: this causes a flash when the nodes move into position. I think the
-    // right fix is to delay the creation of the render plugin until all of the
-    // nodes have been positioned. But I'm not sure if the render plugin is
-    // smart enough to look at existing nodes or it just watches events.
-    for (const [id, nodeModel] of this.mstProgram.nodes) {
-      await this.area.translate(id, { x: nodeModel.x, y: nodeModel.y });
-    }
-  }
-
-  public process = () => {
-    console.log("NodeEditorMST.process");
-
-    // Don't do any processing when we are read-only
-    if (this.readOnly) return;
-
-
-    this.engine.reset();
-
-    // It seems like structures should correctly handle our setup, but from what
-    // I can tell it is reading the empty private nodes and connections from
-    // from our parent. So we make this explicit
-    const graph = structures({nodes: this.getNodes(), connections: this.getConnections()});
-
-    // Because rete engine caches values even if the same node is the
-    // parent of two leaves the data function of that common parent
-    // will only be called once.
-    // debugger;
-    const leafNodes = graph.leaves().nodes();
-    leafNodes.forEach(n => this.engine.fetch(n.id));
-  };
-
-  public logTileChangeEvent = (
-    {operation, change}: Parameters<INodeServices['logTileChangeEvent']>[0]
-  ) => {
-    const logEventName = LogEventName.DATAFLOW_TOOL_CHANGE;
-    logTileChangeEvent(logEventName, { operation, change, tileId: this.tileId });
-  };
-
-  public selectNode = (nodeId: string) => {
-    this.area.emit({ type: "nodepicked", data: { id: nodeId } });
-  };
-
-  public update = (type: "node" | "connection" | "socket" | "control", id: string) => {
-    this.area.update(type, id);
-  };
-
-  public isConnected = (nodeId: string, inputKey: string) => {
-    const matchingConnection = [...this.mstProgram.connections.values()].find(connection =>
-      connection.target === nodeId && connection.targetInput === inputKey);
-
-    return !!matchingConnection;
-  };
-
-  public getOutputVariables = () => {
-    return this.mstContent.outputVariables ?? [];
-  };
-
-  public removeInputConnection = (nodeId: string, inputKey: string) => {
-    const matchingConnection = [...this.mstProgram.connections.values()].find(connection =>
-      connection.target === nodeId && connection.targetInput === inputKey);
-
-    if (!matchingConnection) return;
-    this.removeConnection(matchingConnection.id);
-  };
-
-  public getChannels = () => {
-    return this.mstContent.channels;
-  };
-
-  public async removeNodeAndConnections(nodeId: string) {
-    const node = this.getNode(nodeId);
-    const removedConnections = this.mstProgram.removeNodeAndConnections(nodeId);
-
-    // FIXME: Rete is not happy with this it reports several canceled promises
-
-    removedConnections.forEach(connection => {
-      // Temporary use this approach to get things working
-      this.emit({ type: 'connectionremoved', data: connection });
-    });
-
-    // Temporary use this approach to get things working
-    this.emit({ type: 'noderemoved', data: node});
-
-    // TODO: might need clear the selection, but it might happen automatically
-  }
-
-  private createReteNodeFromNodeModel(id: string, model: IBaseNodeModel) {
-    const nodeTypes: Record<string, NodeClass> =
-    {
-      "Control": ControlNode,
-      "Counter": CounterNode,
-      "Demo Output": DemoOutputNode,
-      "Generator": GeneratorNode,
-      "Live Output": LiveOutputNode,
-      "Logic": LogicNode,
-      "Math": MathNode,
-      "Number": NumberNode,
-      "Sensor": SensorNode,
-      "Timer": TimerNode,
-      "Transform": TransformNode,
-    };
-
-    const constructor = nodeTypes[model.type];
-    if (!constructor) {
-      console.warn(`Can't find constructor for ${model.type}`);
-      return;
-    }
-
-    return new constructor(id, model, this);
-  }
-
-  public async createAndAddNode(nodeType: string, position?: [number, number]) {
-    const id = uniqueId();
-    const newPosition = position ?? getNewNodePosition(this);
-    console.log("createAndAddNode", nodeType, newPosition);
-    this.mstProgram.addNodeSnapshot({
-      id,
-      name: nodeType,
-      x: newPosition[0],
-      y: newPosition[1],
-      data: { type: nodeType }
-    });
-
-    const node = this.getNode(id);
-
-    // Temporarily emit like normal. Ideally the onPatch above would be used
-    // to emit instead, so this is consistent regardless of how the state
-    // is changed.
-    // This is not waiting for the emit before calling the process.
-    // we might need to add it
-    await this.emit({ type: 'nodecreated', data: node });
-
-    this.area.translate(id, {x: newPosition[0], y: newPosition[1]});
-
-    // run the process command so this newly added node can update any controls like the
-    // value control.
-    this.process();
-  }
-
-  disposeNodes() {
-    const nodes = this.getNodes();
-    nodes.forEach(node => (node as IBaseNode).dispose());
-  }
 
   //
   // Methods implementing the Rete `Editor` interface
